Modify missing state verify detector to handle separate method control flow
