+++ findsecbugs-plugin/src/main/java/com/h3xstream/findsecbugs/endpoint/JaxRsEndpointDetector.java
@@ -49,7 +49,7 @@ public class JaxRsEndpointDetector implements Detector {
             for (AnnotationEntry ae : m.getAnnotationEntries()) {
 
                 //Every method mark with @javax.ws.rs.Path is mark as an Endpoint
-                if (ae.getAnnotationType().equals("Ljavax/ws/rs/Path;")) {
+                if (ae.getAnnotationType().equals("Ljavax/ws/rs/Path;") && false) {
                     bugReporter.reportBug(new BugInstance(this, JAXRS_ENDPOINT_TYPE, Priorities.LOW_PRIORITY) //
                             .addClassAndMethod(javaClass, m));
                 }
