+++ findsecbugs-plugin/src/main/java/com/h3xstream/findsecbugs/oidc/state/MissingCheckStateOidcDetector.java
@@ -1,4 +1,4 @@
-package com.h3xstream.findsecbugs.oidc;
+package com.h3xstream.findsecbugs.oidc.state;
 
 import com.h3xstream.findsecbugs.common.matcher.InvokeMatcherBuilder;
 import com.h3xstream.findsecbugs.oidc.data.AnalyzedMethodStateUsage;
@@ -11,38 +11,36 @@ import org.apache.bcel.classfile.JavaClass;
 import org.apache.bcel.classfile.Method;
 import org.apache.bcel.generic.*;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 import static com.h3xstream.findsecbugs.common.matcher.InstructionDSL.invokeInstruction;
 
-public class ImproperStateUsageOidcDetector implements Detector {
+public class MissingCheckStateOidcDetector implements Detector {
+    private final BugReporter bugReporter;
+    private static final String MISSING_VERIFY_OIDC_STATE = "MISSING_VERIFY_OIDC_STATE";
 
-    // Forget comparing state after receiving response to authorization request.
-    // Forget checking state in method reference where state is passed
-
-    private BugReporter bugReporter;
-    private static final String FORGOT_VERIFY_OIDC_STATE = "FORGOT_VERIFY_OIDC_STATE";
-    private static final String POSSIBLY_FORGOT_VERIFY_OIDC_STATE = "POSSIBLY_FORGOT_VERIFY_OIDC_STATE";
-    private static final String POSSIBLY_FORGOT_VERIFY_OIDC_STATE_EXTERNAL_CALL = "POSSIBLY_FORGOT_VERIFY_OIDC_STATE_EXTERNAL_CALL";
     private static final InvokeMatcherBuilder
-            AUTH_REQUEST_INIT = invokeInstruction() //
-                                .atClass("com/nimbusds/openid/connect/sdk/AuthenticationRequest")
-                                .atMethod("<init>");
+            AUTH_RESPONSE_PARSE = invokeInstruction()
+            .atClass("com/nimbusds/openid/connect/sdk/AuthenticationResponseParser")
+            .atMethod("parse");
 
     private static final InvokeMatcherBuilder
             STATE_EQUALS_METHOD = invokeInstruction() //
-                                    .atClass("com/nimbusds/oauth2/sdk/id/State")
-                                    .atMethod("equals")
-                                    .withArgs("(Ljava/lang/Object;)Z");
+            .atClass("com/nimbusds/oauth2/sdk/id/State")
+            .atMethod("equals")
+            .withArgs("(Ljava/lang/Object;)Z");
 
-    public ImproperStateUsageOidcDetector(BugReporter bugReporter) {
+    public MissingCheckStateOidcDetector(BugReporter bugReporter) {
         this.bugReporter = bugReporter;
     }
 
     @Override
     public void visitClassContext(ClassContext classContext) {
         JavaClass javaClass = classContext.getJavaClass();
-        boolean foundAuthRequestWithState = false; // This must be true to trigger the search for state verification
+        boolean foundAuthResponseParse = false; // This must be true to trigger the search for state verification
         boolean foundStateVerify = false; // This must be true in the end to be safe.
         boolean foundStatePassedAsParamToPossibleCheck = false; // We pass state outside of the procedure and need an additional check.
 
@@ -56,7 +54,7 @@ public class ImproperStateUsageOidcDetector implements Detector {
         // Call to a method where state param is called, and caller.
 
         for (Method m : methods) {
-            foundAuthRequestWithState = false;
+            foundAuthResponseParse = false;
             foundStateVerify = false;
             foundStatePassedAsParamToPossibleCheck = false;
 
@@ -71,9 +69,8 @@ public class ImproperStateUsageOidcDetector implements Detector {
                 if (instruction instanceof INVOKESPECIAL) {
                     INVOKESPECIAL invoke = (INVOKESPECIAL) instruction;
                     InvokeInstruction invokeInstruction = (InvokeInstruction) instruction;
-                    if (AUTH_REQUEST_INIT.matches(instruction, cpg) &&
-                            invoke.getSignature(cpg).contains("Lcom/nimbusds/oauth2/sdk/id/State;")) {
-                        foundAuthRequestWithState = true;
+                    if (AUTH_RESPONSE_PARSE.matches(instruction, cpg)) {
+                        foundAuthResponseParse = true;
                     } else if (invoke.getSignature(cpg).contains("Lcom/nimbusds/oauth2/sdk/id/State;") &&
                             // Any other reference where state is passed than AuthenticationResponse assumed to be possible verifier.
                             !invoke.getSignature(cpg).endsWith("Lcom/nimbusds/openid/connect/sdk/AuthenticationRequest;")) {
@@ -85,9 +82,9 @@ public class ImproperStateUsageOidcDetector implements Detector {
                                 MethodAnnotation methodAnnotation = MethodAnnotation.fromXMethod(exactMethod.toXMethod());
                                 methodCallersThatShouldHaveStateCheckForeignMethod.put(
                                         methodAnnotation, new AnalyzedMethodStateUsage(m,
-                                                                            foundAuthRequestWithState,
-                                                                            foundStateVerify,
-                                                                            foundStatePassedAsParamToPossibleCheck
+                                                foundAuthResponseParse,
+                                                foundStateVerify,
+                                                foundStatePassedAsParamToPossibleCheck
                                         ));
                             } catch (ClassNotFoundException e) {
                                 CalledMethodIdentifiers calledMethodIdentifiers = new CalledMethodIdentifiers(
@@ -98,18 +95,18 @@ public class ImproperStateUsageOidcDetector implements Detector {
                                 methodCallersThatShouldHaveStateCheckForeignNotFound.put(
                                         calledMethodIdentifiers,
                                         new AnalyzedMethodStateUsage(m,
-                                                           foundAuthRequestWithState,
-                                                           foundStateVerify,
-                                                           foundStatePassedAsParamToPossibleCheck
+                                                foundAuthResponseParse,
+                                                foundStateVerify,
+                                                foundStatePassedAsParamToPossibleCheck
                                         ));
                             }
                         } else {
                             methodCallersThatShouldHaveStateCheck.put(
                                     calledMethod,
                                     new AnalyzedMethodStateUsage(m,
-                                                        foundAuthRequestWithState,
-                                                        foundStateVerify,
-                                                        foundStatePassedAsParamToPossibleCheck
+                                            foundAuthResponseParse,
+                                            foundStateVerify,
+                                            foundStatePassedAsParamToPossibleCheck
                                     ));
                         }
 
@@ -122,59 +119,60 @@ public class ImproperStateUsageOidcDetector implements Detector {
                 }
             }
 
-            if (foundAuthRequestWithState && !foundStateVerify && !foundStatePassedAsParamToPossibleCheck) {
-                bugReporter.reportBug(new BugInstance(this, FORGOT_VERIFY_OIDC_STATE, Priorities.NORMAL_PRIORITY)
+            if (foundAuthResponseParse && !foundStateVerify && !foundStatePassedAsParamToPossibleCheck) {
+
+                bugReporter.reportBug(new BugInstance(this, MISSING_VERIFY_OIDC_STATE, Priorities.NORMAL_PRIORITY)
                         .addClassAndMethod(javaClass, m));
             }
         }
 
-
-            if(!methodCallersThatShouldHaveStateCheck.isEmpty()) {
-                for(Method localCalledMethod : methodCallersThatShouldHaveStateCheck.keySet()) {
-                    AnalyzedMethodStateUsage analyzedMethod = methodCallersThatShouldHaveStateCheck.get(localCalledMethod);
-                    if(analyzedMethod.foundAuthContextWithState && !analyzedMethod.foundStateVerify && analyzedMethod.foundStatePassedAsParamToPossibleCheck) {
-                        if(!methodsWithStateCheck.contains(localCalledMethod)) {
-                            reportInterproceduralMethodCall(javaClass,
-                                    localCalledMethod,
-                                    analyzedMethod.method);
-
-                        }
-                    }
-                }
-            }
-
-            if(!methodCallersThatShouldHaveStateCheckForeignMethod.isEmpty()) {
-                for(MethodAnnotation calledMethodAnnotation : methodCallersThatShouldHaveStateCheckForeignMethod.keySet()) {
-                    AnalyzedMethodStateUsage analyzedMethod = methodCallersThatShouldHaveStateCheckForeignMethod.get(calledMethodAnnotation);
-                    if(analyzedMethod.foundAuthContextWithState && !analyzedMethod.foundStateVerify && analyzedMethod.foundStatePassedAsParamToPossibleCheck) {
-                        reportInterproceduralMethodCall(
-                                javaClass,
-                                calledMethodAnnotation,
-                                analyzedMethod.method
-                        );
-                    }
-                }
-            }
-
-            if(!methodCallersThatShouldHaveStateCheckForeignNotFound.isEmpty()) {
-                for(CalledMethodIdentifiers calledMethodIdentifiers: methodCallersThatShouldHaveStateCheckForeignNotFound.keySet()) {
-                    AnalyzedMethodStateUsage analyzedMethod = methodCallersThatShouldHaveStateCheckForeignNotFound.get(calledMethodIdentifiers);
-                    if(analyzedMethod.foundAuthContextWithState && !analyzedMethod.foundStateVerify && analyzedMethod.foundStatePassedAsParamToPossibleCheck) {
-                        reportInterproceduralMethodCall(
-                                javaClass,
-                                calledMethodIdentifiers,
+        /*
+        if(!methodCallersThatShouldHaveStateCheck.isEmpty()) {
+            for(Method localCalledMethod : methodCallersThatShouldHaveStateCheck.keySet()) {
+                AnalyzedMethodStateUsage analyzedMethod = methodCallersThatShouldHaveStateCheck.get(localCalledMethod);
+                if(analyzedMethod.foundAuthContextWithState && !analyzedMethod.foundStateVerify && analyzedMethod.foundStatePassedAsParamToPossibleCheck) {
+                    if(!methodsWithStateCheck.contains(localCalledMethod)) {
+                        reportInterproceduralMethodCall(javaClass,
+                                localCalledMethod,
                                 analyzedMethod.method);
+
                     }
                 }
             }
+        }
+
+        if(!methodCallersThatShouldHaveStateCheckForeignMethod.isEmpty()) {
+            for(MethodAnnotation calledMethodAnnotation : methodCallersThatShouldHaveStateCheckForeignMethod.keySet()) {
+                AnalyzedMethodStateUsage analyzedMethod = methodCallersThatShouldHaveStateCheckForeignMethod.get(calledMethodAnnotation);
+                if(analyzedMethod.foundAuthContextWithState && !analyzedMethod.foundStateVerify && analyzedMethod.foundStatePassedAsParamToPossibleCheck) {
+                    reportInterproceduralMethodCall(
+                            javaClass,
+                            calledMethodAnnotation,
+                            analyzedMethod.method
+                    );
+                }
+            }
+        }
+
+        if(!methodCallersThatShouldHaveStateCheckForeignNotFound.isEmpty()) {
+            for(CalledMethodIdentifiers calledMethodIdentifiers: methodCallersThatShouldHaveStateCheckForeignNotFound.keySet()) {
+                AnalyzedMethodStateUsage analyzedMethod = methodCallersThatShouldHaveStateCheckForeignNotFound.get(calledMethodIdentifiers);
+                if(analyzedMethod.foundAuthContextWithState && !analyzedMethod.foundStateVerify && analyzedMethod.foundStatePassedAsParamToPossibleCheck) {
+                    reportInterproceduralMethodCall(
+                            javaClass,
+                            calledMethodIdentifiers,
+                            analyzedMethod.method);
+                }
+            }
+        }
     }
 
     private void reportInterproceduralMethodCall(JavaClass javaClass,
                                                  MethodAnnotation lookupCalledMethod,
                                                  Method callerMethod) {
         bugReporter.reportBug(new BugInstance(this, POSSIBLY_FORGOT_VERIFY_OIDC_STATE_EXTERNAL_CALL, Priorities.LOW_PRIORITY)
-            .addClassAndMethod(javaClass, callerMethod)
-            .addCalledMethod(lookupCalledMethod.toXMethod()));
+                .addClassAndMethod(javaClass, callerMethod)
+                .addCalledMethod(lookupCalledMethod.toXMethod()));
     }
 
     private void reportInterproceduralMethodCall(JavaClass javaClass,
@@ -189,19 +187,19 @@ public class ImproperStateUsageOidcDetector implements Detector {
     private void reportInterproceduralMethodCall(JavaClass javaClass,
                                                  CalledMethodIdentifiers lookupCalledMethodIdentifiers,
                                                  Method callerMethod) {
-            bugReporter.reportBug(new BugInstance(this, POSSIBLY_FORGOT_VERIFY_OIDC_STATE_EXTERNAL_CALL, Priorities.NORMAL_PRIORITY)
-                    .addClassAndMethod(javaClass, callerMethod)
-                    .addString(lookupCalledMethodIdentifiers.toString()));
+        bugReporter.reportBug(new BugInstance(this, POSSIBLY_FORGOT_VERIFY_OIDC_STATE_EXTERNAL_CALL, Priorities.NORMAL_PRIORITY)
+                .addClassAndMethod(javaClass, callerMethod)
+                .addString(lookupCalledMethodIdentifiers.toString())); */
 
     }
 
     private Method findMethodWithName(JavaClass javaClass, String methodName) {
-       for(Method m : javaClass.getMethods()) {
-           if(methodName.equals(m.getName())) {
-               return m;
-           }
-       }
-       return null;
+        for(Method m : javaClass.getMethods()) {
+            if(methodName.equals(m.getName())) {
+                return m;
+            }
+        }
+        return null;
     }
 
     @Override
@@ -209,7 +207,3 @@ public class ImproperStateUsageOidcDetector implements Detector {
 
     }
 }
-
-
-
-
