+++ findsecbugs-samples-java/pom.xml
@@ -87,6 +87,12 @@
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.github.spotbugs</groupId>
+            <artifactId>spotbugs-annotations</artifactId>
+            <version>3.1.12</version>
+            <scope>test</scope>
+        </dependency>
 
     </dependencies>
